+++ test/disqualified/job_test.rb
@@ -6,11 +6,29 @@ class Disqualified::JobTest < ActiveSupport::TestCase
   class OneArgJob
     include Disqualified::Job
 
-    def perform
+    def perform(args)
     end
   end
 
-  test "#perform_async" do
+  test ".job_options" do
+    klass = Class.new do
+      include Disqualified::Job
+      job_options["klass"] = "klass"
+    end
+    llass = Class.new do
+      include Disqualified::Job
+      job_options["llass"] = "llass"
+    end
+    mlass = Class.new(llass) do
+      job_options["mlass"] = "mlass"
+    end
+
+    assert_equal({"klass" => "klass"}, klass.send(:job_options).send(:to_h))
+    assert_equal({"llass" => "llass"}, llass.send(:job_options).send(:to_h))
+    assert_equal({"mlass" => "mlass"}, mlass.send(:job_options).send(:to_h))
+  end
+
+  test ".perform_async" do
     freeze_time do
       assert_difference("Disqualified::Record.count", 1) do
         OneArgJob.perform_async("hello there")
@@ -24,7 +42,7 @@ class Disqualified::JobTest < ActiveSupport::TestCase
     end
   end
 
-  test "#perform_in" do
+  test ".perform_in" do
     freeze_time do
       assert_difference("Disqualified::Record.count", 1) do
         OneArgJob.perform_in(5.minutes, "hello there")
@@ -38,7 +56,7 @@ class Disqualified::JobTest < ActiveSupport::TestCase
     end
   end
 
-  test "#perform_at" do
+  test ".perform_at" do
     assert_difference("Disqualified::Record.count", 1) do
       OneArgJob.perform_at(Time.utc(1970, 1, 1), "hello there")
     end