+++ test/dummy/config/initializers/disqualified.rb
@@ -1,8 +1,8 @@
-Disqualified.configure_server do |config|
+Disqualified.configure do |config|
   config.logger = Rails.logger
   ActiveRecord::Base.logger = Rails.logger
 
-  config.on_error do |error, context|
+  config.on_execution_error do |error, context|
     puts "🔥" * 10
     pp error
     pp context